Change isrunning() semantics to handle non-child process.
Once a server process has started, checking status or stopping it
involves checking if it is running.  The code was using a
os.waitpid() call as part of this test.  However, only a parent
process may wait for a child.  The process trying to stop or
check status isn't the parent, so in these cases the isrunning()
check should not try to wait.